# backend/services-v2/identity-service/openapi.yaml
openapi: 3.1.0
info:
  title: Identity Service API
  version: 2.0.0
  description: Authentication and authorization service for Hospital Management System V2
  contact:
    name: Hospital Management Team
  license:
    name: MIT
servers:
  - url: http://localhost:3021
    description: Development server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/users/{userId}:
    get:
      tags: [User Management]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: OK
    patch:
      tags: [User Management]
      summary: Update user
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName: {type: string}
                phoneNumber: {type: string}
                address: {type: string}
      responses:
        '200':
          description: OK
    delete:
      tags: [User Management]
      summary: Delete user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: OK
  /api/v1/users:
    get:
      tags: [User Management]
      summary: List users (admin only)
      parameters:
        - name: page
          in: query
          schema: {type: integer, default: 1}
        - name: limit
          in: query
          schema: {type: integer, default: 20}
        - name: roleType
          in: query
          schema: {type: string}
        - name: isActive
          in: query
          schema: {type: boolean}
      responses:
        '200':
          description: OK
  /api/v1/users/{userId}/change-password:
    post:
      tags: [User Management]
      summary: Change password
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: {type: string}
                newPassword: {type: string}
      responses:
        '200':
          description: OK
  /api/v1/users/{userId}/lock:
    post:
      tags: [User Management]
      summary: Lock account (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: string}
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: {type: string}
                terminateSessions: {type: boolean, default: true}
      responses:
        '200':
          description: OK
  /api/v1/users/{userId}/unlock:
    post:
      tags: [User Management]
      summary: Unlock account (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: string}
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: {type: string}
      responses:
        '200':
          description: OK
  /api/v1/users/{userId}/assign-role:
    post:
      tags: [User Management]
      summary: Assign role (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roleType]
              properties:
                roleType: {type: string, enum: [admin, doctor, nurse, receptionist, patient]}
                reason: {type: string}
      responses:
        '200':
          description: OK

  # MFA
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT